+++ src/require.d.ts
@@ -1,1 +1,7 @@
 declare function require(s: string): any
+
+
+declare interface String {
+  padStart(length: number, padString: string): string
+}
+