+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
 
 
+- added mark series feature
+  - marked series are displayed first (after all filters are applied)
+- added last queried display
 - we now mark seasons if on episode has changed (added/translated)
 - cleaned up package.json
 - added some filters (show only new, reverse order)